Fix panel quick notifications panel notification dates

+++ src/main/kotlin/com/panomc/platform/route/api/get/panel/PanelQuickNotificationsAndReadAPI.kt
@@ -136,7 +136,7 @@ class PanelQuickNotificationsAndReadAPI : Api() {
                                 mapOf(
                                     "id" to categoryInDB.getInteger(0),
                                     "type_ID" to categoryInDB.getString(2),
-                                    "date" to categoryInDB.getInteger(3),
+                                    "date" to categoryInDB.getLong(3),
                                     "status" to categoryInDB.getString(4),
                                     "isPersonal" to (categoryInDB.getInteger(1) == userID)
                                 )
